improve tests folder structure
